+++ extensions/wubi/CHANGELOG.md
@@ -1,5 +1,8 @@
 # WuBi Input query Changelog
 
-## [Added wubi] - 2025-11-13
+## [Added wubi] - {PR_MERGE_DATE}
+
+- Remove unnecessary items
+- Standardize code
 
 ## [Initial Version] - 2024-08-16